+++ app/cdap/components/FooterPanel/IconStore/ZoomIcon.tsx
@@ -1,0 +1,33 @@
+/*
+ * Copyright © 2022 Cask Data, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+import React from 'react';
+
+export const ZoomIcon = (
+  <svg
+    width="18"
+    height="19"
+    viewBox="0 0 18 19"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    id="zoom-icon"
+  >
+    <path
+      d="M12.7549 11.7559H11.9649L11.6849 11.4859C12.6649 10.3459 13.2549 8.86586 13.2549 7.25586C13.2549 3.66586 10.3449 0.755859 6.75488 0.755859C3.16488 0.755859 0.254883 3.66586 0.254883 7.25586C0.254883 10.8459 3.16488 13.7559 6.75488 13.7559C8.36488 13.7559 9.84488 13.1659 10.9849 12.1859L11.2549 12.4659V13.2559L16.2549 18.2459L17.7449 16.7559L12.7549 11.7559ZM6.75488 11.7559C4.26488 11.7559 2.25488 9.74586 2.25488 7.25586C2.25488 4.76586 4.26488 2.75586 6.75488 2.75586C9.24488 2.75586 11.2549 4.76586 11.2549 7.25586C11.2549 9.74586 9.24488 11.7559 6.75488 11.7559ZM7.25488 4.75586H6.25488V6.75586H4.25488V7.75586H6.25488V9.75586H7.25488V7.75586H9.25488V6.75586H7.25488V4.75586Z"
+      fill="#5F6368"
+    />
+  </svg>
+);
